Update spell mistake of package name
