+++ src/main/java/com/dhl/virtual/model/Employee.java
@@ -1,0 +1,32 @@
+package com.dhl.virtual.model;
+
+public class Employee {
+    private String name;
+    private String department;
+    private double salary;
+
+    // Constructor
+    public Employee(String name, String department, double salary) {
+        this.name = name;
+        this.department = department;
+        this.salary = salary;
+    }
+
+    // Getters
+    public String getName() {
+        return name;
+    }
+
+    public String getDepartment() {
+        return department;
+    }
+
+    public double getSalary() {
+        return salary;
+    }
+
+    @Override
+    public String toString() {
+        return "Employee{name='" + name + "', department='" + department + "', salary=" + salary + "}";
+    }
+}
