+++ src/main/java/com/dhl/virtual/model/NotificationMessage.java
@@ -1,0 +1,50 @@
+package com.dhl.virtual.model;
+
+public class NotificationMessage {
+    private String userId;
+    private String type;
+    private String message;
+    private String timestamp;
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public String getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(String timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    @Override
+    public String toString() {
+        return "NotificationMessage{" +
+                "userId='" + userId + '\'' +
+                ", type='" + type + '\'' +
+                ", message='" + message + '\'' +
+                ", timestamp='" + timestamp + '\'' +
+                '}';
+    }
+}
