+++ src/main/java/com/dhl/virtual/redish_service/RedishCacheService.java
@@ -1,0 +1,6 @@
+package com.dhl.virtual.redish_service;
+
+public interface RedishCacheService {
+
+}
+
