--- src/main/java/com/dhl/virtusa/model/Employee.java
@@ -1,32 +1,0 @@
-package com.dhl.virtusa.model;
-
-public class Employee {
-    private String name;
-    private String department;
-    private double salary;
-
-    // Constructor
-    public Employee(String name, String department, double salary) {
-        this.name = name;
-        this.department = department;
-        this.salary = salary;
-    }
-
-    // Getters
-    public String getName() {
-        return name;
-    }
-
-    public String getDepartment() {
-        return department;
-    }
-
-    public double getSalary() {
-        return salary;
-    }
-
-    @Override
-    public String toString() {
-        return "Employee{name='" + name + "', department='" + department + "', salary=" + salary + "}";
-    }
-}
