--- src/main/java/com/dhl/virtusa/redish_service/RedishCacheService.java
@@ -1,6 +1,0 @@
-package com.dhl.virtusa.redish_service;
-
-public interface RedishCacheService {
-
-}
-
