+++ src/test/java/com/dhl/virtual/FeaturePrepUtilTest.java
@@ -1,0 +1,41 @@
+package com.dhl.virtual;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class FeaturePrepUtilTest {
+
+    @Test
+    void getInstance() {
+
+    }
+
+    @Test
+    void findLongestString() {
+    }
+
+    @Test
+    void findAvgOfGivenNumber() {
+    }
+
+    @Test
+    void findSumOfSquareOfEven() {
+    }
+
+    /*@BeforeAll
+    void initialize() {
+        // This method can be used to set up any required state before tests run
+        // For example, initializing static variables or loading configurations
+    }
+
+    @AfterAll
+    void afterAll() {
+        // This method can be used to clean up any state after all tests have run
+        // For example, resetting static variables or closing resources
+    }*/
+
+    //Interface first test
+    //Builder design pattern test
+    //Singleton design pattern test
+}
