+++ src/test/java/com/dhl/virtual/controller/KafkaControllerTest.java
@@ -1,0 +1,14 @@
+package com.dhl.virtual.controller;
+
+import org.junit.jupiter.api.Test;
+
+class KafkaControllerTest {
+
+    @Test
+    void kafkaPage() {
+    }
+
+    @Test
+    void sendMessage() {
+    }
+}
