+++ src/test/java/com/dhl/virtual/controller/NotificationControllerTest.java
@@ -1,0 +1,14 @@
+package com.dhl.virtual.controller;
+
+import org.junit.jupiter.api.Test;
+
+class NotificationControllerTest {
+
+    @Test
+    void notifyUser() {
+    }
+
+    @Test
+    void getNotifications() {
+    }
+}
