+++ src/test/java/com/dhl/virtual/error_handle/ErrorHandlingDeserializerTest.java
@@ -1,0 +1,18 @@
+package com.dhl.virtual.error_handle;
+
+import org.junit.jupiter.api.Test;
+
+class ErrorHandlingDeserializerTest {
+
+    @Test
+    void configure() {
+    }
+
+    @Test
+    void deserialize() {
+    }
+
+    @Test
+    void close() {
+    }
+}
