--- src/test/java/com/dhl/virtusa/controller/KafkaControllerTest.java
@@ -1,16 +1,0 @@
-package com.dhl.virtusa.controller;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class KafkaControllerTest {
-
-    @Test
-    void kafkaPage() {
-    }
-
-    @Test
-    void sendMessage() {
-    }
-}
