--- src/test/java/com/dhl/virtusa/controller/NotificationControllerTest.java
@@ -1,16 +1,0 @@
-package com.dhl.virtusa.controller;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class NotificationControllerTest {
-
-    @Test
-    void notifyUser() {
-    }
-
-    @Test
-    void getNotifications() {
-    }
-}
