--- src/test/java/com/dhl/virtusa/error_handle/ErrorHandlingDeserializerTest.java
@@ -1,20 +1,0 @@
-package com.dhl.virtusa.error_handle;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ErrorHandlingDeserializerTest {
-
-    @Test
-    void configure() {
-    }
-
-    @Test
-    void deserialize() {
-    }
-
-    @Test
-    void close() {
-    }
-}
